Удаляет избыточные модули
