+++ js/main.js
@@ -1,6 +1,3 @@
-import { getSettings } from './data.js';
-import { createPhoto } from './create-photo.js';
+import { generatePhotosList } from './data.js';
 
-const generatePhotosList = (length) => Array.from({ length }, createPhoto);
-
-generatePhotosList(getSettings().DESCRIPTION_MAX_COUNT);
+generatePhotosList();
